test(con-005): port the rate-limit tests onto OpcodeRateLimiter
`dig-peer-protocol` 0.5.0 replaces chia's `ProtocolMessageTypes`-keyed
`RateLimiter` with `OpcodeRateLimiter`, keyed by the raw wire opcode, so
the CON-005 tests move onto it: `OpcodeRateLimits::from(&limits)` re-keys
the same chia table the old fixtures used, and `allow` replaces
`handle_message`.

The one semantic difference — `admit` does not charge a refused message,
where chia's inbound limiter did — changes no assertion here: every test
exhausts its budget with ADMITTED frames and then asserts the next one is
refused, so nothing depends on a rejection consuming budget. The
charge-on-reject property still lives in dig-gossip's own
`DigRateLimiter`, which keeps the `incoming` flag.

`test_rate_limit_blocks_oversized_message` gets stronger rather than
merely ported: the bound is pinned from both sides (at `max_size` admits,
one byte over refuses) and the refusal is asserted as `Unsendable`, not
just "not admitted" — a size refusal survives every window roll, so
reporting it as `Deferred` would send a retrying caller to wait on a
budget that can never clear. `Admission` is re-exported to make that
assertable.

Also clears the warnings the previously-failing target masked: unused
imports across seven test files and two more no-op `as u8` casts.

+++ src/service/holdings_announce.rs
@@ -1191,7 +1191,7 @@ mod tests {
     fn frame_and_lift_round_trip() {
         let a = sample();
         let msg = frame_holdings_announce(&a);
-        assert_eq!(msg.msg_type as u8, HOLDINGS_ANNOUNCE);
+        assert_eq!(msg.msg_type, HOLDINGS_ANNOUNCE);
         assert_eq!(msg.id, None);
         assert_eq!(holdings_announce_payload(&msg), Some(a));
     }
